Updated program structure
