+++ Parsing_Program/Parsing_Program.pyproj
@@ -22,6 +22,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Combiner.py" />
+    <Compile Include="Excel_Functions.py" />
     <Compile Include="Parsing_Program.py" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
